re-implement calculation of fuel consumption
