+++ Isochrone/router.py
@@ -6,7 +6,7 @@ import numpy as np
 import datetime as dt
 
 import utils
-from isochrone import RoutingAlg
+from RoutingAlgTimeMin import RoutingAlgTimeMin
 from polars import Boat
 from routeparams import RouteParams
 from weather import wind_function
@@ -40,13 +40,13 @@ def modified_isochrone_routing(start, #r_la1, r_lo1
                 iso (Isochrone) - next isochrone
     """
 
-    ra=RoutingAlg(start,finish,start_time)
+    ra=RoutingAlgTimeMin(start,finish,start_time)
     ra.set_steps(steps)
     ra.set_pruning_settings(params['ISOCHRONE_PRUNE_SECTOR_DEG_HALF'],params['ISOCHRONE_PRUNE_SEGMENTS'])
     ra.set_variant_segments(params['ROUTER_HDGS_SEGMENTS'], params ['ROUTER_HDGS_INCREMENTS_DEG'])
     min_time_route=ra.recursive_routing(boat, winds, delta_time,  verbose)
 
-    #min_time_route.print_route()
+    min_time_route.print_route()
     return min_time_route
